% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raeddat.R
\name{RAEDDA_transductive}
\alias{RAEDDA_transductive}
\title{Robust Eigenvalues Decomposition Discriminant Analysis (transductive approach)}
\usage{
RAEDDA_transductive(Xtrain, cltrain, Xtest, modelscope = NULL, G = NULL,
  alpha_Xtrain = 0, alpha_Xtest = 0.05, tol = 10^-5, iterlim = 1000,
  Aitken = TRUE, ...)
}
\arguments{
\item{Xtrain}{A numeric matrix of observations where rows correspond to observations and columns correspond to variables. The group membership of each observation is known - labeled data.}

\item{cltrain}{A vector (if numeric it will be coerced to factor) with distinct entries representing a classification of the corresponding observations in Xtrain.}

\item{Xtest}{A numeric matrix of observations where rows correspond to observations and columns correspond to variables. The group membership of each observation may not be known - unlabeled data.}

\item{modelscope}{A character string indicating the desired models to be tested. With default NULL, all available models are tested. The models available for univariate and multivariate data are described in \code{\link{modelvec}}}

\item{G}{A numeric vector indicating the number of expected classes in Xtest. With default NULL, models with G = length(unique(cltrain)): (length(unique(cltrain))+3) are tested.}

\item{alpha_Xtrain}{The proportion of observations to be trimmed in Xtrain.}

\item{alpha_Xtest}{The proportion of observations to be trimmed in Xtest.}

\item{tol}{A non-negative number, with default 10^-5, which is a measure of how strictly convergence is defined.}

\item{iterlim}{A non-negative integer, with default 1000, which is the desired limit on the maximum number of iterations.}

\item{Aitken}{A logical value with default TRUE which tests for convergence using Aitken acceleration. If value is set to FALSE, convergence is tested by comparing tol to the change in log-likelihood between two consecutive iterations. For further information on Aitken acceleration, see \code{\link{Aitken}}}

\item{...}{Arguments passed to or from other methods}
}
\value{
An object of class "raeddat" providing a list of output components for each model in modelscope, with the Best model (according to BIC) first
}
\description{
Robust Eigenvalues Decomposition Discriminant Analysis (transductive approach)
}
